+++ realm/realm-library/src/main/cpp/io_realm_internal_TestUtil.cpp
@@ -32,45 +32,15 @@ throwOrGetExpectedMessage(JNIEnv *env, jlong testcase, bool should_throw)
             if (should_throw)
                 ThrowException(env, ClassNotFound, "parm1", "parm2");
             break;
-        case NoSuchField:
-            expect = "java.lang.NoSuchFieldException: Field 'parm2' could not be located in class io.realm.parm1";
-            if (should_throw)
-                ThrowException(env, NoSuchField, "parm1", "parm2");
-            break;
-        case NoSuchMethod:
-            expect = "java.lang.NoSuchMethodException: Method 'parm2' could not be located in class io.realm.parm1";
-            if (should_throw)
-                ThrowException(env, NoSuchMethod, "parm1", "parm2");
-            break;
         case IllegalArgument:
             expect = "java.lang.IllegalArgumentException: Illegal Argument: parm1";
             if (should_throw)
                 ThrowException(env, IllegalArgument, "parm1", "parm2");
             break;
-        case IOFailed:
-            expect = "io.realm.exceptions.RealmIOException: Failed to open parm1. parm2";
-            if (should_throw)
-                ThrowException(env, IOFailed, "parm1", "parm2");
-            break;
-        case FileNotFound:
-            expect = "io.realm.exceptions.RealmIOException: File not found: parm1.";
-            if (should_throw)
-                ThrowException(env, FileNotFound, "parm1", "parm2");
-            break;
-        case FileAccessError:
-            expect = "io.realm.exceptions.RealmIOException: Failed to access: parm1. parm2";
-            if (should_throw)
-                ThrowException(env, FileAccessError, "parm1", "parm2");
-            break;
         case IndexOutOfBounds:
             expect = "java.lang.ArrayIndexOutOfBoundsException: parm1";
             if (should_throw)
                 ThrowException(env, IndexOutOfBounds, "parm1", "parm2");
-            break;
-        case TableInvalid:
-            expect = "java.lang.IllegalStateException: Illegal State: parm1";
-            if (should_throw)
-                ThrowException(env, TableInvalid, "parm1", "parm2");
             break;
         case UnsupportedOperation:
             expect = "java.lang.UnsupportedOperationException: parm1";
@@ -92,31 +62,19 @@ throwOrGetExpectedMessage(JNIEnv *env, jlong testcase, bool should_throw)
             if (should_throw)
                 ThrowException(env, RuntimeError, "parm1", "parm2");
             break;
-        case RowInvalid:
-            expect = "java.lang.IllegalStateException: Illegal State: parm1";
-            if (should_throw)
-                ThrowException(env, RowInvalid, "parm1", "parm2");
-            break;
-        case CrossTableLink:
-            expect = "java.lang.IllegalStateException: This class is referenced by other classes. Remove those fields first before removing this class.";
-            if (should_throw)
-                ThrowException(env, CrossTableLink, "parm1");
-            break;
         case BadVersion:
             expect = "io.realm.internal.async.BadVersionException: parm1";
             if (should_throw)
                 ThrowException(env, BadVersion, "parm1", "parm2");
-            break;
-        case LockFileError:
-            expect = "io.realm.exceptions.IncompatibleLockFileException: parm1";
-            if (should_throw)
-                ThrowException(env, LockFileError, "parm1", "parm2");
             break;
         case IllegalState:
             expect = "java.lang.IllegalStateException: parm1";
             if (should_throw)
                 ThrowException(env, IllegalState, "parm1");
             break;
+        // FIXME: This is difficult to test right now. Need to refactor the test.
+        // See https://github.com/realm/realm-java/issues/3348
+        // case RealmFileError:
         default:
             break;
     }